requirements.txt
asgiref==3.6.0
Django==4.0.8
django-cors-headers==3.10.1
djangorestframework==3.13.1
pytz==2022.7.1
sqlparse==0.4.3
